Extract AuthScreen entrance and tab-switch animations into useAuthFormAnimation hook
