+++ src/frontend/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useButtonAnimation';
+export * from './useAuthFormAnimation';
